Moving folders. Better naming

+++ mepinta/core/python_core/mepinta/plugins_creation/import_cpp/PluginCPPImplementationImporter.py
@@ -24,20 +24,16 @@ from mepinta.plugins_creation.PluginPackageCreator import PluginPackageCreator
 from mepinta.plugins_creation.base import PluginCreatorBase
 from common.path import joinPath
 
-#TODO: Add new data types to the data types database (done?)
+# TODO: Add new data types to the data types database (done?)
 class PluginCPPImplementationImporter(PluginCreatorBase):
   def __post_init__(self):
     self.plugin_creator = PluginPackageCreator(context=self.context)
     self.default_impl = "Debug"
 
   def __createSymlink(self, src, dst, force=True):
-    try:
-      if force:
-        os.remove(dst)
-    except:
-      pass
-    self.context.log.debug("Linking %r <- %r"%(src, dst))
-    #os.__createSymlink(src, dst)
+    if self._pathExists(dst) and force:
+      os.remove(dst)
+    self.context.log.debug("Linking %r <- %r" % (src, dst))
     os.symlink(src, dst)
 
   def __projectSrcPath(self, plugin_fdirs):
@@ -54,60 +50,60 @@ class PluginCPPImplementationImporter(PluginCreatorBase):
       return file_path.endswith(".c")   or file_path.endswith(".h") or file_path.endswith(".cpp") or file_path.endswith(".hpp")
 
   def _ignoredFiles(self, plugin_fdirs):
-    return plugin_fdirs.build_names+['.cproject','.project']
+    return plugin_fdirs.build_names + ['.cproject', '.project']
 
-  def _linkCPPToPackage(self,src_path, plugin_fdirs):
+  def _linkCPPFiles(self, src_path, plugin_fdirs):
     for file_path in os.listdir(src_path):
       if self._acceptedExtension(file_path):
-        self.context.log.verbose("Linking %r "%file_path)
-        self.__createSymlink(joinPath(src_path, file_path), joinPath(plugin_fdirs.getPath("code"),file_path))
+        self.context.log.verbose("Linking %r " % file_path)
+        self.__createSymlink(joinPath(src_path, file_path), joinPath(plugin_fdirs.getPath("code"), file_path))
       elif file_path not in self._ignoredFiles(plugin_fdirs):
-        self.context.log.warning("File or folder %r will not be synchronized. (in path %r)"%(file_path, src_path))
-  
+        self.context.log.warning("File or folder %r will not be synchronized. (in path %r)" % (file_path, src_path))
+
   def _linkSharedLibraries(self, plugin_fdirs):
     build_paths = plugin_fdirs.implementationLibs()
     for build, path in build_paths.items():
-      self.context.log.verbose("Linking %r."%build)
-      self.__createSymlink(path, plugin_fdirs.getPath("module")+".%s.so.implementation"%build)
+      self.context.log.verbose("Linking %r." % build)
+      self.__createSymlink(path, plugin_fdirs.getPath("module") + ".%s.so.implementation" % build)
     if self.default_impl in build_paths:
-      self.context.log.verbose("Linking %r as default."%self.default_impl)
-      self.__createSymlink(build_paths[self.default_impl], plugin_fdirs.getPath("module")+".so.implementation")
+      self.context.log.verbose("Linking %r as default." % self.default_impl)
+      self.__createSymlink(build_paths[self.default_impl], plugin_fdirs.getPath("module") + ".so.implementation")
 
   def _linkEclipseFiles(self, plugin_fdirs):
     src_path = self.__projectSrcPath(plugin_fdirs)
     if self._pathExists(src_path):
-      self._linkCPPToPackage(src_path, plugin_fdirs)
+      self._linkCPPFiles(src_path, plugin_fdirs)
       self._linkSharedLibraries(plugin_fdirs)
     else:
-      self.context.log.warning("Eclipse path %r doesn't exist."%src_path)
+      self.context.log.warning("Eclipse path %r doesn't exist." % src_path)
       return False
     return True
 
   def _linkIncludeFiles(self, plugin_fdirs):
-    #is it a data_type plugin? then do de link.
+    # is it a data_type plugin? then do de link.
     if plugin_fdirs.lang_group == "data_types":
       src_path = self.__projectSrcPath(plugin_fdirs)
       for file_path in os.listdir(src_path):
         if self._acceptedExtension(file_path, header_only=True):
-          self.context.log.verbose("Linking %r "%file_path)
-          self.__createSymlink(joinPath(src_path, file_path), joinPath(plugin_fdirs.getPath("data_types_include"),file_path))
-    
+          self.context.log.verbose("Linking %r " % file_path)
+          self.__createSymlink(joinPath(src_path, file_path), joinPath(plugin_fdirs.getPath("data_types_include"), file_path))
+
   def importPlugin(self, plugin_fdirs):
     ''' Checks for existence of files in the independent project directory (Eclipse or QtCreator).
       Creates corresponding directories if they don't exist.
-      Links files in the created directories. 
+      Links files in the created directories.
         * .c, .cpp, .h and .hpp in the code directory
         * .so to .so.implementation in the plugin directory
       Creates the plugin manifest if not present.
-      
+
       plugin_fdirs: provides all the configuration of the plugin.
         Where to get the files and where to send them.
     '''
     if self._linkEclipseFiles(plugin_fdirs):
-      #if it was able to create the plugin module in the mepinta tree
-      #then link the include files for data type plugins
+      # if it was able to create the plugin module in the mepinta tree
+      # then link the include files for data type plugins
       self._linkIncludeFiles(plugin_fdirs)
-    
-        
+
+
 if __name__ == "__main__":
   pass
